requirements.txt
rstr
asyncpg
py-cord
nest-asyncio
python-dotenv
python-dateutil